Working on performance
Looked at storing the object in the ViewBean vs recreat in xthml call.
Also removed the prepare call to the top.. it only needs to be called
once.  in the end.. all our calls can be prepared when the system starts
up and we simply reference them through out the code.  vs preparing each
time